clean up controller servers to be in serperate files
