+++ catkin_ws/src/controls/src/servers/BaseServer.py
@@ -1,10 +1,8 @@
 #! /usr/bin/env python3
 
 import rospy
-import actionlib
 from geometry_msgs.msg import Point
 from geometry_msgs.msg import Pose
-from auv_msgs.msg import SuperimposerAction, SuperimposerFeedback, SuperimposerGoal, SuperimposerResult, StateAction, StateFeedback, StateResult
 from std_msgs.msg import Float64, Bool
 import time
 
@@ -140,165 +138,3 @@ class BaseServer():
             self.pub_theta_y_pid.publish(Float64(rotation.y))
         if(self.goal.do_theta_z.data):
             self.pub_theta_z_pid.publish(Float64(rotation.z))
-
-
-class StateServer(BaseServer):
-    def __init__(self):
-        super().__init__()
-        self.establish_pid_publishers()
-        self.establish_pid_enable_publishers()
-        self.establish_state_subscribers()
-        self.server = actionlib.SimpleActionServer('state_server', StateAction, execute_cb= self.callback, auto_start = False)
-        self.server.start()
-
-    def callback(self, goal):
-        print("got a message")
-        # set the PIDs
-        #print(goal.pose)
-        self.goal = goal
-        self.enable_pids(goal)
-        if(goal.displace.data):
-            goal_position, goal_rotation = self.dispalce_goal(goal)
-        else:
-            goal_position, goal_rotation = goal.position, goal.rotation
-        print(goal_position,goal_rotation)
-        self.publish_setpoints(goal_position,goal_rotation)
-
-        # monitor when reached pose
-        self.wait_for_settled(goal_position,goal_rotation)
-
-        # rospy.loginfo("Succeeded")
-        if(not self.cancelled):
-            self.server.set_succeeded()
-    
-    def dispalce_goal(self, goal):
-        goal_x = goal.position.x + self.position.x
-        goal_y = goal.position.y + self.position.y
-        goal_z = goal.position.z + self.position.z
-
-        goal_theta_x = goal.rotation.x + self.theta_x
-        goal_theta_y = goal.rotation.y + self.theta_y
-        goal_theta_z = goal.rotation.z + self.theta_z
-
-        goal_position = Point()
-        goal_position.x = goal_x
-        goal_position.y = goal_y
-        goal_position.z = goal_z
-
-        goal_rotation = Point()
-        goal_rotation.x = goal_theta_x
-        goal_rotation.y = goal_theta_y
-        goal_rotation.z = goal_theta_z
-
-        return goal_position, goal_rotation
-
-
-class SuperimposerServer(BaseServer):
-    def __init__(self):
-        super().__init__()
-        self.establish_pid_enable_publishers()
-        self.establish_pid_publishers()
-        self.establish_state_subscribers()
-
-        self.pub_x_effort = None
-        self.pub_y_effort = None
-        self.pub_z_effort = None
-        self.pub_theta_x_effort = None
-        self.pub_theta_y_effort = None
-        self.pub_theta_z_effort = None
-
-    
-    def displace_goal(self,displace):
-        print("displacing goal")
-        if(self.goal == None):
-            return displace
-
-        new_goal = displace
-        new_goal.effort.force.x += self.goal.effort.force.x
-        new_goal.effort.force.y += self.goal.effort.force.y
-        new_goal.effort.force.z += self.goal.effort.force.z
-        new_goal.effort.torque.x += self.goal.effort.torque.x
-        new_goal.effort.torque.y += self.goal.effort.torque.y
-        new_goal.effort.torque.z += self.goal.effort.torque.z
-        return new_goal
-    
-    def cancel(self):
-        if(self.goal.do_surge.data):
-            self.pub_x_effort.publish(0)
-        if(self.goal.do_sway.data):
-            self.pub_y_effort.publish(0)
-        if(self.goal.do_heave.data):
-            self.pub_z_effort.publish(0)
-        if(self.goal.do_roll.data):
-            self.pub_theta_x_effort.publish(0)
-        if(self.goal.do_pitch.data):
-            self.pub_theta_y_effort.publish(0)
-        if(self.goal.do_yaw.data):
-            self.pub_theta_z_effort.publish(0)
-
-    
-    def callback(self, goal):
-        print("received new goal")
-        print(goal)
-        if(goal.displace.data):
-            self.goal = self.displace_goal(goal)
-        else:
-            self.goal = goal
-        #unset pids
-        if(goal.do_surge.data):
-            self.pub_x_enable.publish(Bool(False))
-            self.pub_x_effort.publish(self.goal.effort.force.x)
-        if(goal.do_sway.data):
-            #unset pids
-            self.pub_y_enable.publish(Bool(False))
-            self.pub_y_effort.publish(self.goal.effort.force.y)
-        if(goal.do_heave.data):
-            #unset pids
-            self.pub_z_enable.publish(Bool(False))
-            self.pub_z_effort.publish(self.goal.effort.force.z)
-        if(goal.do_roll.data):
-            #unset pids
-            self.pub_theta_x_enable.publish(Bool(False))
-            self.pub_theta_x_effort.publish(self.goal.effort.torque.x)
-        if(goal.do_pitch.data):
-            #unset pids
-            self.pub_theta_y_enable.publish(Bool(False))
-            self.pub_theta_y_effort.publish(self.goal.effort.torque.y)
-        if(goal.do_yaw.data):
-            #unset pids
-            self.pub_theta_z_enable.publish(Bool(False))
-            self.pub_theta_z_effort.publish(self.goal.effort.torque.z)
-
-        self.server.set_succeeded()
-
-
-
-class LocalSuperimposerServer(SuperimposerServer):
-    def __init__(self):
-        super().__init__()
-        self.server = actionlib.SimpleActionServer('superimposer_local_server', SuperimposerAction, execute_cb= self.callback, auto_start = False)
-
-
-        self.pub_x_effort = rospy.Publisher('surge', Float64, queue_size=50)
-        self.pub_y_effort = rospy.Publisher('sway', Float64, queue_size=50)
-        self.pub_z_effort = rospy.Publisher('heave', Float64, queue_size=50)
-        self.pub_theta_x_effort = rospy.Publisher('roll', Float64, queue_size=50)
-        self.pub_theta_y_effort = rospy.Publisher('pitch', Float64, queue_size=50)
-        self.pub_theta_z_effort = rospy.Publisher('yaw', Float64, queue_size=50)
-  
-        self.server.start()
-
-class GlobalSuperimposerServer(SuperimposerServer):
-    def __init__(self):
-        super().__init__()
-        self.server = actionlib.SimpleActionServer('superimposer_global_server', SuperimposerAction, execute_cb= self.callback, auto_start = False)
-
-        self.pub_x_effort = rospy.Publisher('global_x', Float64, queue_size=50)
-        self.pub_y_effort = rospy.Publisher('global_y', Float64, queue_size=50)
-        self.pub_z_effort = rospy.Publisher('global_z', Float64, queue_size=50)
-        self.pub_theta_x_effort = rospy.Publisher('roll', Float64, queue_size=50)
-        self.pub_theta_y_effort = rospy.Publisher('pitch', Float64, queue_size=50)
-        self.pub_theta_z_effort = rospy.Publisher('yaw', Float64, queue_size=50)
-
-        self.server.start()
-
